requirements.txt
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.1
matplotlib==3.4.3
seaborn==0.11.2
pytest==7.3.1
pyyaml==6.0